gcc code fixes

+++ ACG-Assignment3/src/NanoFlannKdPMRP.hpp
@@ -190,13 +190,13 @@ private:
 
 				/* Determine transmitted ray direction for refraction */
 				if (into)
-					tdir = (origDir * nnt - normal * (ddn * nnt + sqrt(cos2t)));
+					tdir = (origDir * nnt - normal * (ddn * nnt + std::sqrt(cos2t)));
 				else
-					tdir = (origDir * nnt + normal * (ddn * nnt + sqrt(cos2t)));
+					tdir = (origDir * nnt + normal * (ddn * nnt + std::sqrt(cos2t)));
 
 				tdir = glm::normalize(tdir);
 
-				/* Determine R0 for Schlick´s approximation */
+				/* Determine R0 for Schlickï¿½s approximation */
 				float a = nt - nc;
 				float b = nt + nc;
 				float R0 = a*a / (b*b);
@@ -208,7 +208,7 @@ private:
 				else
 					c = 1 - glm::dot(tdir, normal);
 
-				/* Compute Schlick´s approximation of Fresnel equation */
+				/* Compute Schlickï¿½s approximation of Fresnel equation */
 				float Re = R0 + (1.f - R0) *c*c*c*c*c;   /* Reflectance */
 				float Tr = 1.f - Re;                     /* Transmittance */
 
@@ -327,7 +327,7 @@ private:
 
 		if (depth > 5 || !p)   /* After 5 bounces or if max reflectivity is zero */
 		{
-			if (drand48() < p*0.9)            /* Russian Roulette */
+			if (drand48() < p * 0.9)            /* Russian Roulette */
 				col = col * (1.0f / p);        /* Scale estimator to remain unbiased */
 			else
 			{
@@ -344,9 +344,9 @@ private:
 				nl = nl*-1.0f;
 
 			/* Compute random reflection vector on hemisphere */
-			auto r1 = 2.0f * PI * drand48();
-			auto r2 = drand48();
-			auto r2s = sqrt(r2);
+			auto r1 = static_cast<float>(2.0 * PI * drand48());
+			auto r2 = static_cast<float>(drand48());
+			auto r2s = std::sqrt(r2);
 
 			/* Set up local orthogonal coordinate system u,v,w on surface */
 			auto w = nl;
@@ -360,9 +360,9 @@ private:
 			auto v = glm::cross(w, u);
 
 			/* Random reflection vector d */
-			auto d = u * cos(r1) * r2s +
-				v * sin(r1) * r2s +
-				w * sqrt(1.f - r2);
+			auto d = u * std::cos(r1) * r2s +
+				v * std::sin(r1) * r2s +
+				w * std::sqrt(1.f - r2);
 
 			return emission + col * GetPhotonMapRadianceEstimate(intersectionInfo);
 		}
@@ -415,13 +415,13 @@ private:
 
 			/* Determine transmitted ray direction for refraction */
 			if (into)
-				tdir = (origDir * nnt - normal * (ddn * nnt + sqrt(cos2t)));
+				tdir = (origDir * nnt - normal * (ddn * nnt + std::sqrt(cos2t)));
 			else
-				tdir = (origDir * nnt + normal * (ddn * nnt + sqrt(cos2t)));
+				tdir = (origDir * nnt + normal * (ddn * nnt + std::sqrt(cos2t)));
 
 			tdir = glm::normalize(tdir);
 
-			/* Determine R0 for Schlick´s approximation */
+			/* Determine R0 for Schlickï¿½s approximation */
 			auto a = nt - nc;
 			auto b = nt + nc;
 			auto R0 = a*a / (b*b);
@@ -433,7 +433,7 @@ private:
 			else
 				c = 1 - glm::dot(tdir, normal);
 
-			/* Compute Schlick´s approximation of Fresnel equation */
+			/* Compute Schlickï¿½s approximation of Fresnel equation */
 			auto Re = R0 + (1.f - R0) *c*c*c*c*c;   /* Reflectance */
 			auto Tr = 1.f - Re;                     /* Transmittance */
